@@ -1,49 +1,69 @@
-source "https://rubygems.org"
+source 'https://rubygems.org'
 
-gem "rails", "3.2.13"
 
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem "sass-rails",   "~> 3.2.3"
-  gem "coffee-rails", "~> 3.2.1"
-  gem "uglifier", ">= 1.0.3"
-  gem "jquery-rails"
-  gem "jquery-ui-rails"
-  gem "jquery-modal-rails"
-  gem "bootstrap-datepicker-rails"
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
+gem 'rails', '>= 5.0.0.beta2', '< 5.1'
+# Use postgreSQL as the database for Active Record
+gem 'pg'
+# Use Puma as the app server
+gem 'puma'
+# Use SCSS for stylesheets
+gem 'sass-rails', '~> 5.0'
+# Use Uglifier as compressor for JavaScript assets
+gem 'uglifier', '>= 1.3.0'
+# Use CoffeeScript for .coffee assets and views
+gem 'coffee-rails', '~> 4.1.0'
+# See https://github.com/rails/execjs#readme for more supported runtimes
+# gem 'therubyracer', platforms: :ruby
+
+# Use jquery as the JavaScript library
+gem 'jquery-rails'
+gem "jquery-ui-rails", "~> 4.2.1"
+gem "jquery-modal-rails"
+gem "bootstrap-datepicker-rails"
+#autocomplete
+gem "rails3-jquery-autocomplete"
+gem "wysiwyg-rails"
+#notification
+gem "gritter"
+
+# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
+gem 'turbolinks'
+# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
+gem 'jbuilder', '~> 2.0'
+# Action Cable dependencies for the Redis adapter
+gem 'redis', '~> 3.0'
+# Use ActiveModel has_secure_password
+# gem 'bcrypt', '~> 3.1.7'
+
+# Use Capistrano for deployment
+# gem 'capistrano-rails', group: :development
+
+group :development, :test do
+  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
+  gem 'byebug'
 end
 
 group :development do
-  gem "rails_layout"
+  # Access an IRB console on exception pages or by using <%= console %> in views
+  gem 'web-console', '~> 3.0'
 end
 
-gem "devise"
+# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
+gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
 
-#autocomplete
-gem "rails3-jquery-autocomplete"
-
-#notification
-gem "gritter", "1.0.3"
+gem "devise", '>= 4.0.0.rc1', '< 4.1'
 gem "awesome_nested_set"
-
 #To use tagging
 gem "acts-as-taggable-on"
-
 # To use Geolocations
 gem "geocoder"
 
-# To use ActiveModel has_secure_password
-gem "bcrypt-ruby", "~> 3.0.0"
-
-gem "pg"
-
 # Search Engine
-gem "sunspot_rails"
+#gem "sunspot_rails"
 
 gem "magnific-popup-rails"
 gem "slim-rails"
-gem "simple_form"
 gem "country_select"
 gem "journey"
 gem "arel"
@@ -54,7 +74,3 @@ gem "truncate_html"
 gem "carrierwave"
 gem "rmagick"
 gem "mini_magick"
-
-#chat
-gem 'private_pub'
-gem 'thin'
