@@ -1,15 +1,14 @@
 class ApplicationController < ActionController::Base
   include ActionView::Helpers::TextHelper
   protect_from_forgery
-  force_ssl
+  #force_ssl
 
   helper :all # include all helpers, all the time
   helper_method :my_pins
-  helper_method :init_chats
 
   private
 
-  before_filter :set_locale
+  before_action :set_locale
 
   def auth_user
     redirect_to new_user_registration_url unless user_signed_in?
@@ -21,12 +20,6 @@ class ApplicationController < ActionController::Base
 
   def my_pins
     @my_pins = Pin.where(:user_id => @current_user.id).to_a
-  end
-
-  def init_chats
-    if user_signed_in?
-      @init_chats = Conversation.involving(@current_user)
-    end
   end
 
   def mobile_browser?
