@@ -1,5 +1,5 @@
 class DashboardController < ApplicationController
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
   def index
     @user = User.find(current_user.id)
     @skills = @user.tag_counts_on(:skills)
