@@ -1,3 +1,3 @@
 class Users::SessionsController < Devise::SessionsController
-  skip_before_filter :require_no_authentication, :only => [:new, :create]
+  skip_before_action :require_no_authentication, :only => [:new, :create]
 end
