@@ -1,5 +1,4 @@
-class PinTag < ActiveRecord::Base
-  attr_accessible :pin_id, :tag_id
+class PinTag < ApplicationRecord
   belongs_to :pin
   belongs_to :tag
 end
